simplify by removing builder
